+++ dlz.web/src/main/java/com/dlz/app/uim/bean/UserRole.java
@@ -10,18 +10,18 @@ public class UserRole implements java.io.Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 5070462418563344534L;
-	private Integer member_id;
-	private Integer role_id;
-	public Integer getMember_id() {
+	private Long member_id;
+	private Long role_id;
+	public Long getMember_id() {
 		return member_id;
 	}
-	public void setMember_id(Integer member_id) {
+	public void setMember_id(Long member_id) {
 		this.member_id = member_id;
 	}
-	public Integer getRole_id() {
+	public Long getRole_id() {
 		return role_id;
 	}
-	public void setRole_id(Integer role_id) {
+	public void setRole_id(Long role_id) {
 		this.role_id = role_id;
 	}
 }
