+++ dlz.web/src/main/java/com/dlz/app/uim/service/IUserRoleInfoService.java
@@ -6,7 +6,7 @@ import com.dlz.framework.bean.JSONMap;
 import com.dlz.framework.db.modal.ResultMap;
 
 public interface IUserRoleInfoService {
-	public boolean saveRoleInfo(int id,JSONMap info);
-	public JSONMap getRoleInfo(Integer id);
-	public List<ResultMap> getRoleInfo(List<Integer> ids);
+	public boolean saveRoleInfo(Long id,JSONMap info);
+	public JSONMap getRoleInfo(Long id);
+	public List<ResultMap> getRoleInfo(List<Long> ids);
 }
